refactor: new archive archive service architecture
