+++ .github/workflows/pr.yml
@@ -36,7 +36,7 @@ jobs:
       - name: Setup Protocol Buffer Compiler
         uses: arduino/setup-protoc@v3
         with:
-          version: "29.3"
+          version: '29.3'
 
       - name: Cache go-generate-fast
         uses: actions/cache@v4
@@ -55,10 +55,10 @@ jobs:
       - name: golangci-lint
         uses: golangci/golangci-lint-action@v8
         env:
-          CGO_ENABLED: "1"
-          CGO_CFLAGS: "-I${{ github.workspace }}/libs"
-          CGO_LDFLAGS: "-L${{ github.workspace }}/libs -lcodex -Wl,-rpath,${{ github.workspace }}/libs"
-          LD_LIBRARY_PATH: "${{ github.workspace }}/libs:${{ env.LD_LIBRARY_PATH }}"
+          CGO_ENABLED: '1'
+          CGO_CFLAGS: '-I${{ github.workspace }}/libs'
+          CGO_LDFLAGS: '-L${{ github.workspace }}/libs -lstorage -Wl,-rpath,${{ github.workspace }}/libs'
+          LD_LIBRARY_PATH: '${{ github.workspace }}/libs:${{ env.LD_LIBRARY_PATH }}'
         with:
           version: v2.3.1
           args: --build-tags=gowaku_no_rln,lint
