+++ .vscode/settings.json
@@ -1,5 +1,5 @@
 {
-  "go.testTags": "gowaku_skip_migrations,gowaku_no_rln,use_codex",
+  "go.testTags": "gowaku_skip_migrations,gowaku_no_rln,use_logos_storage",
   "cSpell.words": ["unmarshalling"],
   "gopls": {
     "formatting.local": "github.com/status-im/status-go",
@@ -15,9 +15,11 @@
   },
   "go.testEnvVars": {
     "CGO_CFLAGS": "-I${workspaceFolder}/libs",
-    "CGO_LDFLAGS": "-L${workspaceFolder}/libs -lcodex -Wl,-rpath,${workspaceFolder}/libs"
+    "CGO_LDFLAGS": "-L${workspaceFolder}/libs -lstorage -Wl,-rpath,${workspaceFolder}/libs"
   },
   "go.buildFlags": ["-tags=gowaku_skip_migrations,gowaku_no_rln"],
-  "go.testFlags": ["-tags=gowaku_skip_migrations,gowaku_no_rln,use_codex"],
+  "go.testFlags": [
+    "-tags=gowaku_skip_migrations,gowaku_no_rln,use_logos_storage"
+  ],
   "idf.pythonInstallPath": "/usr/bin/python"
 }
