+++ protocol/communities/archive/archive_service.go
@@ -1,0 +1,57 @@
+package archive
+
+//go:generate go tool mockgen -package=mock_archive -source=archive_service.go -destination=mock/archive/archive_service.go
+
+import (
+	"context"
+	"time"
+
+	"github.com/status-im/status-go/crypto/types"
+	messagingtypes "github.com/status-im/status-go/messaging/types"
+	archivetypes "github.com/status-im/status-go/protocol/communities/archive/types"
+	"github.com/status-im/status-go/protocol/protobuf"
+)
+
+type ArchiveServiceBackend interface {
+	SetOnline(bool)
+	Start() error
+	Stop() error
+	IsStarted() bool
+	IsReady() bool
+	SeedHistoryArchive(communityID types.HexBytes, archiveLink string) error
+	UnseedHistoryArchive(communityID types.HexBytes, archiveLink string)
+	IsSeedingHistoryArchive(communityID types.HexBytes, archiveLink string) bool
+	DownloadHistoryArchives(communityID types.HexBytes, archiveLink string, cancelTask chan struct{}) (*archivetypes.HistoryArchiveDownloadTaskInfo, error)
+	CreateHistoryArchiveFromMessages(communityID types.HexBytes, messages []*messagingtypes.ReceivedMessage, topics []messagingtypes.ContentTopic, startDate time.Time, endDate time.Time, partition time.Duration, encrypt bool) ([]string, error)
+	CreateHistoryArchiveFromDB(communityID types.HexBytes, topics []messagingtypes.ContentTopic, startDate time.Time, endDate time.Time, partition time.Duration, encrypt bool) ([]string, error)
+	CreateAndSeedHistoryArchive(communityID types.HexBytes, topics []messagingtypes.ContentTopic, startDate time.Time, endDate time.Time, partition time.Duration, encrypt bool) error
+	LoadArchiveMessages(ctx context.Context, communityID types.HexBytes, archiveLink string, downloadedArchiveID string) ([]*protobuf.WakuMessage, error)
+
+	// GetHistoryArchiveLink(communityID types.HexBytes) (string, error)
+	// Those will be private methods in the concrete implementations
+	// ExtractMessagesFromHistoryArchive(communityID types.HexBytes, archiveLink string, logosStorageIndex *protobuf.LogosStorageWakuMessageArchiveIndex) ([]*protobuf.WakuMessage, error)
+	// LoadHistoryArchiveIndex(ctx context.Context, myKey *ecdsa.PrivateKey, communityID types.HexBytes, indexCid string, isLocal bool) (*protobuf.LogosStorageWakuMessageArchiveIndex, error)
+}
+
+type ArchiveService interface {
+	// ArchiveServiceBackend provides a proxy interface to the underlying storage-specific
+	// implementation of the archive service.
+	ArchiveServiceBackend
+	// Storage-agnostic operations
+	GetHistoryArchiveLink(communityID types.HexBytes) (string, error)
+	GetCommunityChatsFilters(communityID types.HexBytes) (messagingtypes.ChatFilters, error)
+	GetCommunityChatsTopics(communityID types.HexBytes) ([]messagingtypes.ContentTopic, error)
+	GetHistoryArchivePartitionStartTimestamp(communityID types.HexBytes) (uint64, error)
+	CreateAndSeedHistoryArchive(communityID types.HexBytes, topics []messagingtypes.ContentTopic, startDate time.Time, endDate time.Time, partition time.Duration, encrypt bool) error
+	// StartHistoryArchiveTasksInterval(community *Community, interval time.Duration)
+	StartHistoryArchiveTasksInterval(communityID types.HexBytes, chatID string, encrypted bool, interval time.Duration)
+	StopHistoryArchiveTasksInterval(communityID types.HexBytes)
+	GetHistoryArchiveDownloadTask(communityID string) *archivetypes.HistoryArchiveDownloadTask
+	AddHistoryArchiveDownloadTask(communityID string, task *archivetypes.HistoryArchiveDownloadTask)
+	RemoveHistoryArchiveDownloadTask(communityID string)
+	PublishHistoryArchivesSeedingSignal(communityID types.HexBytes)
+	GetDownloadedMessageArchiveIDs(communityID types.HexBytes) ([]string, error)
+	SaveMessageArchiveID(communityID types.HexBytes, hash string) error
+	GetMessageArchiveIDsToImport(communityID types.HexBytes) ([]string, error)
+	SetMessageArchiveIDImported(communityID types.HexBytes, hash string, imported bool) error
+}
