+++ protocol/communities/archive/commons/archive_commons.go
@@ -1,0 +1,5 @@
+package commons
+
+import "errors"
+
+var ErrArchiveTimedout = errors.New("archive has timed out")
