+++ protocol/communities/archive/consts/archive_consts.go
@@ -1,0 +1,3 @@
+package consts
+
+const MaxArchiveSizeInBytes = 30000000
