--- protocol/communities/manager_archive_file_nop.go
@@ -1,59 +1,0 @@
-//go:build disable_torrent
-// +build disable_torrent
-
-package communities
-
-import (
-	"crypto/ecdsa"
-	"time"
-
-	"github.com/status-im/status-go/crypto/types"
-	messagingtypes "github.com/status-im/status-go/messaging/types"
-	"github.com/status-im/status-go/protocol/protobuf"
-)
-
-type ArchiveFileManagerNop struct{}
-
-func (amm *ArchiveFileManagerNop) CreateHistoryArchiveTorrentFromMessages(communityID types.HexBytes, messages []*messagingtypes.ReceivedMessage, topics []messagingtypes.ContentTopic, startDate time.Time, endDate time.Time, partition time.Duration, encrypt bool) ([]string, error) {
-	return nil, nil
-}
-
-func (amm *ArchiveFileManagerNop) CreateHistoryArchiveTorrentFromDB(communityID types.HexBytes, topics []messagingtypes.ContentTopic, startDate time.Time, endDate time.Time, partition time.Duration, encrypt bool) ([]string, error) {
-	return nil, nil
-}
-
-func (amm *ArchiveFileManagerNop) CreateHistoryArchiveLogosStorageFromMessages(communityID types.HexBytes, messages []*messagingtypes.ReceivedMessage, topics []messagingtypes.ContentTopic, startDate time.Time, endDate time.Time, partition time.Duration, encrypt bool) ([]string, error) {
-	return nil, nil
-}
-
-func (amm *ArchiveFileManagerNop) CreateHistoryArchiveLogosStorageFromDB(communityID types.HexBytes, topics []messagingtypes.ContentTopic, startDate time.Time, endDate time.Time, partition time.Duration, encrypt bool) ([]string, error) {
-	return nil, nil
-}
-
-func (amm *ArchiveFileManagerNop) SaveMessageArchiveID(communityID types.HexBytes, hash string) error {
-	return nil
-}
-
-func (amm *ArchiveFileManagerNop) GetMessageArchiveIDsToImport(communityID types.HexBytes) ([]string, error) {
-	return nil, nil
-}
-
-func (amm *ArchiveFileManagerNop) SetMessageArchiveIDImported(communityID types.HexBytes, hash string, imported bool) error {
-	return nil
-}
-
-func (amm *ArchiveFileManagerNop) ExtractMessagesFromHistoryArchive(communityID types.HexBytes, archiveID string) ([]*protobuf.WakuMessage, error) {
-	return nil, nil
-}
-
-func (amm *ArchiveFileManagerNop) ExtractMessagesFromLogosStorageHistoryArchive(communityID types.HexBytes, archiveID string) ([]*protobuf.WakuMessage, error) {
-	return nil, nil
-}
-
-func (amm *ArchiveFileManagerNop) GetHistoryArchiveMagnetlink(communityID types.HexBytes) (string, error) {
-	return "", nil
-}
-
-func (amm *ArchiveFileManagerNop) LoadHistoryArchiveIndexFromFile(myKey *ecdsa.PrivateKey, communityID types.HexBytes) (*protobuf.WakuMessageArchiveIndex, error) {
-	return nil, nil
-}
