+++ protocol/messenger.go
@@ -42,6 +42,8 @@ import (
 	"github.com/status-im/status-go/multiaccounts/settings"
 	"github.com/status-im/status-go/protocol/common"
 	"github.com/status-im/status-go/protocol/communities"
+	"github.com/status-im/status-go/protocol/communities/archive"
+	archivetypes "github.com/status-im/status-go/protocol/communities/archive/types"
 	"github.com/status-im/status-go/protocol/ens"
 	"github.com/status-im/status-go/protocol/identity/alias"
 	"github.com/status-im/status-go/protocol/identity/identicon"
@@ -101,7 +103,7 @@ type Messenger struct {
 	pushNotificationClient    *pushnotificationclient.Client
 	pushNotificationServer    PushNotificationServer
 	communitiesManager        *communities.Manager
-	archiveManager            communities.ArchiveService
+	archiveManager            archive.ArchiveService
 	communitiesKeyDistributor communities.KeyDistributor
 	accountsManager           AccountsManager
 	mentionsManager           *MentionManager
@@ -365,7 +367,7 @@ func NewMessenger(
 		return nil, err
 	}
 
-	amc := &communities.ArchiveManagerConfig{
+	amc := &archivetypes.ArchiveManagerConfig{
 		TorrentConfig:      c.torrentConfig,
 		LogosStorageConfig: c.logosStorageConfig,
 		Logger:             logger,
@@ -378,7 +380,7 @@ func NewMessenger(
 	// Depending on the OS go will choose whether to use the "communities/manager_archive_nop.go" or
 	// "communities/manager_archive.go" version of this function based on the build instructions for those files.
 	// See those file for more details.
-	archiveManager := communities.NewArchiveManager(amc)
+	archiveManager := archive.NewArchiveManager(amc)
 
 	settings, err := accounts.NewDB(database)
 	if err != nil {
@@ -490,6 +492,26 @@ func NewMessenger(
 	}
 
 	return messenger, nil
+}
+
+func (m *Messenger) SetupArchiveManager(amc *archivetypes.ArchiveManagerConfig) {
+	if (amc.Logger == nil) {
+		amc.Logger = m.logger
+	}
+	if (amc.Persistence == nil) {
+		amc.Persistence = m.communitiesManager.GetPersistence()
+	}
+	if (amc.Messaging == nil) {
+		amc.Messaging = m.messaging
+	}
+	if (amc.Identity == nil) {
+		amc.Identity = m.identity
+	}
+	if (amc.Publisher == nil) {
+		amc.Publisher = m.communitiesManager
+	}
+
+	m.archiveManager = archive.NewArchiveManager(amc)
 }
 
 func (m *Messenger) processSentMessage(id string) error {
@@ -752,9 +774,7 @@ func (m *Messenger) handleConnectionChange(online bool) {
 	}
 
 	// Update torrent manager
-	if m.archiveManager != nil {
-		m.archiveManager.SetOnline(online)
-	}
+	m.archiveManager.SetOnline(online)
 
 	// Publish contact code
 	if online && m.shouldPublishContactCode {
