+++ protocol/messenger_testing_utils.go
@@ -60,7 +60,7 @@ func (m *MessengerSignalsHandlerMock) HistoryArchivesProtocolDisabled()         
 func (m *MessengerSignalsHandlerMock) CreatingHistoryArchives(string)            {}
 func (m *MessengerSignalsHandlerMock) NoHistoryArchivesCreated(string, int, int) {}
 func (m *MessengerSignalsHandlerMock) HistoryArchivesCreated(string, int, int)   {}
-func (m *MessengerSignalsHandlerMock) HistoryArchivesSeeding(string, bool, bool) {}
+func (m *MessengerSignalsHandlerMock) HistoryArchivesSeeding(string)             {}
 func (m *MessengerSignalsHandlerMock) HistoryArchivesUnseeded(string)            {}
 func (m *MessengerSignalsHandlerMock) ManifestFetched(string, string)            {}
 func (m *MessengerSignalsHandlerMock) IndexDownloadCompleted(string, string)     {}
