+++ protocol/protobuf/application_metadata_message.proto
@@ -50,7 +50,7 @@ message ApplicationMetadataMessage {
     SYNC_BOOKMARK = 40;
     SYNC_CLEAR_HISTORY = 41;
     SYNC_SETTING = 42;
-    COMMUNITY_MESSAGE_ARCHIVE_MAGNETLINK = 43;
+    COMMUNITY_MESSAGE_ARCHIVE_LINK = 43;
     SYNC_PROFILE_PICTURES = 44;
     SYNC_ACCOUNT = 45;
     ACCEPT_CONTACT_REQUEST = 46;
@@ -97,6 +97,5 @@ message ApplicationMetadataMessage {
     COMMUNITY_SHARED_ADDRESSES_REQUEST = 89;
     COMMUNITY_SHARED_ADDRESSES_RESPONSE = 90;
     BACKED_UP_MESSAGE_BATCH = 91;
-    COMMUNITY_MESSAGE_ARCHIVE_INDEX_CID = 92;
   }
 }
