+++ protocol/protobuf/communities.proto
@@ -112,7 +112,7 @@ message CommunityDescription {
   map<string,CommunityChat> chats = 6;
   repeated string ban_list = 7 [deprecated = true];
   map<string,CommunityCategory> categories = 8;
-  uint64 archive_magnetlink_clock = 9;
+  uint64 archive_link_clock = 9;
   CommunityAdminSettings admin_settings = 10;
   string intro_message = 11;
   string outro_message = 12;
@@ -125,7 +125,6 @@ message CommunityDescription {
   map<string,CommunityBanInfo>banned_members = 19;
   // request to resend revealed addresses
   uint64 resend_accounts_clock = 20;
-  uint64 archive_index_cid_clock = 21;
   // key is hash ratchet key_id + seq_no
   map<string, bytes> privateData = 100;
 }
@@ -204,11 +203,10 @@ message CommunityRequestToJoinResponse {
   bool accepted = 3;
   bytes grant = 4;
   bytes community_id = 5;
-  string magnet_uri = 6;
+  string archive_link = 6;
   reserved 7, 8;
   // CommunityDescription protocol message with owner signature
   bytes community_description_protocol_message = 9;
-  string index_cid = 10;
 }
 
 message CommunityRequestToLeave {
@@ -216,14 +214,9 @@ message CommunityRequestToLeave {
   bytes community_id = 2;
 }
 
-message CommunityMessageArchiveMagnetlink {
+message CommunityMessageArchiveLink {
   uint64 clock = 1;
-  string magnet_uri = 2;
-}
-
-message CommunityMessageArchiveIndexCid {
-  uint64 clock = 1;
-  string cid = 2;
+  string archive_link = 2;
 }
 
 message WakuMessage {
