+++ signal/events_community_archives.go
@@ -62,8 +62,6 @@ type HistoryArchivesCreatedSignal struct {
 
 type HistoryArchivesSeedingSignal struct {
 	CommunityID string `json:"communityId"`
-	MagnetLink  bool   `json:"magnetLink"`
-	IndexCid    bool   `json:"indexCid"`
 }
 
 type HistoryArchivesUnseededSignal struct {
@@ -126,11 +124,9 @@ func SendHistoryArchivesCreated(communityID string, from int, to int) {
 	})
 }
 
-func SendHistoryArchivesSeeding(communityID string, magnetLink bool, indexCid bool) {
+func SendHistoryArchivesSeeding(communityID string) {
 	send(EventHistoryArchivesSeeding, HistoryArchivesSeedingSignal{
 		CommunityID: communityID,
-		MagnetLink:  magnetLink,
-		IndexCid:    indexCid,
 	})
 }
